SessionId Long to Int

+++ microtask/src/main/kotlin/logic/TaskController.kt
@@ -59,22 +59,23 @@ class TaskController private constructor(private val ws: WSTaskServer,
         // How to send something that is not a Task?
 
         //the following is for the websocket android test
-        //val message = PayloadWrapper(Services.instanceId().toLong(),
+        //val message = PayloadWrapper(Services.instanceId()
+        //,
         //        WSOperations.ADD_LEADER, "ok")
         if (members.isNotEmpty()) {
-            val message = PayloadWrapper(Services.instanceId().toLong(),
+            val message = PayloadWrapper(Services.instanceId(),
                     WSOperations.LIST_MEMBERS, MembersAdditionNotification(members.keys().toList()).toJson())
             ws.sendMessage(leader.second, message)
             //leader.second.remote.sendString(members.keys().toList().toJson())
         }
         //ws.sendMessage(leader.second, message)
-        ws.sendMessage(leader.second, PayloadWrapper(Services.instanceId().toLong(),WSOperations.LEADER_RESPONSE,"ok"))
+        ws.sendMessage(leader.second, PayloadWrapper(Services.instanceId(),WSOperations.LEADER_RESPONSE,"ok"))
     }
 
     fun addMember(member: Member, session: Session) {
         members[member] = session
         if (leader.second.isOpen) {
-            val message = PayloadWrapper(Services.instanceId().toLong(),
+            val message = PayloadWrapper(Services.instanceId(),
                     WSOperations.ADD_MEMBER, MembersAdditionNotification(members.keys().toList()).toJson())
             ws.sendMessage(leader.second, message)
         }
@@ -83,7 +84,7 @@ class TaskController private constructor(private val ws: WSTaskServer,
     fun addTask(task: Task, member: Member) {
         if (members.containsKey(member)) {
             taskMemberAssociationList.add(TaskMemberAssociation.create(task, member))
-            val message = PayloadWrapper(Services.instanceId().toLong(),
+            val message = PayloadWrapper(Services.instanceId(),
                     WSOperations.ADD_TASK, TaskAssignment(member, task).toJson())
             ws.sendMessage(members[member]!!, message)
             "$dbUrl/task/add".httpPost().body(task.toJson()).responseString()
@@ -95,13 +96,13 @@ class TaskController private constructor(private val ws: WSTaskServer,
         with(taskMemberAssociationList.firstOrNull { it.task.id == task.id }) {
             this?.let {
                 taskMemberAssociationList.remove(this)
-                val message = PayloadWrapper(Services.instanceId().toLong(),
+                val message = PayloadWrapper(Services.instanceId(),
                         WSOperations.REMOVE_TASK, TaskAssignment(member, task).toJson())
                 ws.sendMessage(members[member]!!, message)
                 "$dbUrl/task/${it.task.id}".httpDelete().responseString()
                 "$visorUrl/remove/${it.task.id}".httpDelete().responseString()
             }
-                    ?: ws.sendMessage(leader.second, PayloadWrapper(Services.instanceId().toLong(),
+                    ?: ws.sendMessage(leader.second, PayloadWrapper(Services.instanceId(),
                             WSOperations.ERROR_REMOVING_TASK, TaskError(task, "").toJson()))
         }
     }
@@ -110,12 +111,12 @@ class TaskController private constructor(private val ws: WSTaskServer,
         with(taskMemberAssociationList.firstOrNull { it.task.id == task.id }) {
             this?.let {
                 it.task.statusId = task.statusId
-                val message = PayloadWrapper(Services.instanceId().toLong(),
+                val message = PayloadWrapper(Services.instanceId(),
                         WSOperations.CHANGE_TASK_STATUS, TaskAssignment(member, this.task).toJson())
                 ws.sendMessage(members[member]!!, message)
                 ws.sendMessage(leader.second, message)
                 "$dbUrl/task/${it.task.id}/status/${it.task.statusId}".httpPut().responseString()
-            } ?: ws.sendMessage(session, PayloadWrapper(Services.instanceId().toLong(),
+            } ?: ws.sendMessage(session, PayloadWrapper(Services.instanceId(),
                     WSOperations.ERROR_CHANGING_STATUS, StatusError(task.statusId, task, "").toJson()))
         }
     }
@@ -123,7 +124,7 @@ class TaskController private constructor(private val ws: WSTaskServer,
     fun getAllActivities(member: Member, session: Session) {
         leader = Pair(member, session)
         if (activityList.isNotEmpty()) {
-            val message = PayloadWrapper(Services.instanceId().toLong(),
+            val message = PayloadWrapper(Services.instanceId(),
                     WSOperations.SET_ALL_ACTIVITIES, ActivityAdditionNotification(activityList).toJson())
             ws.sendMessage(leader.second, message)
         }
